Wordpress API theme
This theme only contains the minimum codebase necesary to run WP REST API.For this project, I create custom Routes just to get more control of the necessary content. In this project, we need 3 arrays: ths first one is sticky post, the second is the featured posts and the last one is newest posts.

I created two routes, one for the loading of the first content and the second route just to pagination.